+++ installer/CorumIntegrationInstaller.Core/AppConstants.cs
@@ -10,11 +10,9 @@ public static class AppConstants
     public const string GeometryDashProcessName = "GeometryDash";
     public const string SteamAppId = "322170";
     public const string GeodeInstallUrl = "https://geode-sdk.org/install";
+    public const string ReleaseAssetName = ModId + ".geode";
     public const long MaximumPackageBytes = 256L * 1024L * 1024L;
 
     public static Uri LatestReleaseApiUri { get; } =
         new($"https://api.github.com/repos/{RepositoryOwner}/{RepositoryName}/releases/latest");
-
-    public static string GetReleaseAssetName(string normalizedVersion) =>
-        $"Corum-Integration-{normalizedVersion}.geode";
 }